+++ gazebo/common/HeightmapData.hh
@@ -15,11 +15,14 @@
  *
 */
 
-#ifndef _GAZEBO_HEIGHTMAPDATA_HH_
-#define _GAZEBO_HEIGHTMAPDATA_HH_
+#ifndef GAZEBO_COMMON_HEIGHTMAPDATA_HH_
+#define GAZEBO_COMMON_HEIGHTMAPDATA_HH_
 
+#include <string>
 #include <vector>
 #include <ignition/math/Vector3.hh>
+
+#include "gazebo/gazebo_config.h"
 
 #include "gazebo/common/CommonTypes.hh"
 #include "gazebo/math/Vector3.hh"
@@ -65,6 +68,32 @@ namespace gazebo
       /// \brief Get the maximum terrain's elevation.
       /// \return The maximum terrain's elevation.
       public: virtual float GetMaxElevation() const = 0;
+    };
+
+    /// \class HeightmapDataLoader HeightmapData.hh common/common.hh
+    /// \brief Helper class for loading heightmap data.
+    class GZ_COMMON_VISIBLE HeightmapDataLoader
+    {
+      /// \brief Load a terrain file specified by _filename. The terrain file
+      /// format might be an image or a DEM file. libgdal is required to enable
+      /// DEM support. For a list of all raster formats supported you can type
+      /// the command "gdalinfo --formats".
+      /// \param[in] _filename The path to the terrain file.
+      /// \return 0 when the operation succeeds to load a file or -1 when fails.
+      public: static HeightmapData *LoadTerrainFile(
+          const std::string &_filename);
+
+      /// \brief Load a DEM specified by _filename as a terrain file.
+      /// \param[in] _filename The path to the terrain file.
+      /// \return 0 when the operation succeeds to load a file or -1 when fails.
+      private: static HeightmapData *LoadDEMAsTerrain(
+          const std::string &_filename);
+
+      /// \brief Load an image specified by _filename as a terrain file.
+      /// \param[in] _filename The path to the terrain file.
+      /// \return 0 when the operation succeeds to load a file or -1 when fails.
+      private: static HeightmapData *LoadImageAsTerrain(
+          const std::string &_filename);
     };
     /// \}
   }